+++ lib/model/android_battery_info.dart
@@ -7,21 +7,30 @@ class AndroidBatteryInfo {
   int chargeTimeRemaining = -1;
   String health = "unknown";
   String pluggedStatus = "unknown";
+  String technology = "unknown";
+  int batteryLevel;
+  int remainingEnergy = -1;
+  int scale;
   int temperature = -1;
   int voltage = -1;
-  int batteryLevel;
+  bool present = true;
   ChargingStatus chargingStatus;
 
-  AndroidBatteryInfo(
-      {this.chargeTimeRemaining,
-      this.currentAverage,
-      this.currentNow,
-      this.health,
-      this.pluggedStatus,
-      this.temperature,
-      this.voltage,
-      this.batteryLevel,
-      this.chargingStatus});
+  AndroidBatteryInfo({
+    this.batteryLevel,
+    this.chargeTimeRemaining,
+    this.chargingStatus,
+    this.currentAverage,
+    this.currentNow,
+    this.health,
+    this.pluggedStatus,
+    this.present,
+    this.remainingEnergy,
+    this.scale,
+    this.technology,
+    this.temperature,
+    this.voltage,
+  });
 
   /// Serialise data back to json from the model
   Map<String, dynamic> toJson() {
@@ -33,11 +42,16 @@ class AndroidBatteryInfo {
     data["currentNow"] = this.currentNow;
     data["health"] = this.health;
     data["pluggedStatus"] = this.pluggedStatus;
+    data["present"] = this.present;
+    data["scale"] = this.scale;
+    data["remainingEnergy"] = this.remainingEnergy;
     data["temperature"] = this.temperature;
+    data["technology"] = this.technology;
     data["voltage"] = this.voltage;
     return data;
   }
 
+  /// Retrieves the chargin status from the native value
   ChargingStatus getChargingStatus(String status) {
     switch (status) {
       case "charging":
@@ -51,7 +65,7 @@ class AndroidBatteryInfo {
     }
   }
 
-  @override
+  /// Deserialize data from json
   AndroidBatteryInfo.fromJson(Map<String, dynamic> json) {
     this.batteryLevel = json["batteryLevel"];
     this.chargingStatus = getChargingStatus(json["chargingStatus"]);
@@ -60,6 +74,10 @@ class AndroidBatteryInfo {
     this.currentNow = json["currentNow"];
     this.health = json["health"];
     this.pluggedStatus = json["pluggedStatus"];
+    this.present = json["present"];
+    this.scale = json["scale"];
+    this.remainingEnergy = json["remainingEnergy"];
+    this.technology = json["technology"];
     this.temperature = json["temperature"];
     this.voltage = json["voltage"];
   }