+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 ## 1.0.2
 
 * Fixed bug: remaining charge time was always returning -1
+
+## 1.0.3
+
+* Added more android battery information:
+    - Technology (e.g Li-ion)
+    - Scale (maximum scale of the battery)
+    - If battery is present
+    - Remaining energy (nanowatt-hours)